Dev to Mian merging new features and updations

+++ src/application/useCases/common/s3.useCase.ts
@@ -1,26 +1,29 @@
-import { randomUUID } from "crypto";
 import { awsConfig } from "../../../config/env";
-import { log } from "../../../shared/logger/logger";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
+import { BadRequestError } from "../../../shared/error/appError";
+import { toAppError } from "../../../shared/error/handleUnknownError";
 import { DeleteObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import { ISignedUrlService } from "../../../domain/interfaces/services/ISignedUrl.service";
-import { CreareFileUploadPresignedUrlRequest, CreareFileUploadPresignedUrlResponse, CreateFileSignedUrlRequest } from "../../dtos/common.dto";
+import { CreateFileUploadPresignedUrlOutput, CreateFileSignedUrlInput, CreateFileUploadPresignedUrlInput } from "../../dtos/common.dto";
+import { generateId } from "../../../shared/utils/generateId";
+import { IdType } from "../../../shared/utils/types";
 
 export class CreateFileUploadPresignedUrlUseCase {
     constructor(
         private s3Client: S3Client
     ) { };
 
-    async execute(data: CreareFileUploadPresignedUrlRequest): Promise<CreareFileUploadPresignedUrlResponse> {
+    async execute(input: CreateFileUploadPresignedUrlInput): Promise<CreateFileUploadPresignedUrlOutput> {
         try {
-            const { fileName, fileType, folderName } = data;
-
-            if (!fileName || !fileType || !folderName) throw new Error("Invalid request");
+            const { fileName, fileType, folderName } = input;
+            if (!fileName || !fileType || !folderName) {
+                throw new BadRequestError();
+            }
 
             const ext = fileName.split(".").pop();
-            if (!ext) throw new Error("Invalid file name");
+            if (!ext) throw new BadRequestError();
 
-            const key = `${folderName}/${Date.now()}-${randomUUID()}.${ext}`;
+            const key = `${folderName}/${Date.now()}-${generateId({ type: IdType.FILE })}.${ext}`;
 
             const command = new PutObjectCommand({
                 Bucket: awsConfig.awsS3BucketName,
@@ -34,9 +37,8 @@ export class CreateFileUploadPresignedUrlUseCase {
                 key,
                 uploadUrl
             };
-        } catch (error) {
-            log.error("CreateFileUploadPresignedUrlUseCase failed", error as Error);
-            throw error;
+        } catch (error: unknown) {
+            throw toAppError(error, "Failed to create file upload presigned url");
         };
     };
 };
@@ -46,18 +48,18 @@ export class CreateFileSignedUrlUseCase {
         private signedUrlService: ISignedUrlService,
     ) { };
 
-    async execute(payload: CreateFileSignedUrlRequest): Promise<string> {
+    async execute(input: CreateFileSignedUrlInput): Promise<string> {
         try {
-            const { key } = payload;
-            if (!key) throw new Error("Noe key found");
+            const { key } = input;
+            if (!key) {
+                throw new BadRequestError();
+            }
 
             const signedUrl = await this.signedUrlService.get(key);
-            if (!signedUrl) throw new Error("Failed to generate signed url");
 
             return signedUrl;
-        } catch (error) {
-            log.error("CreateFileSignedUrlUseCase failed", error as Error);
-            throw error;
+        } catch (error: unknown) {
+            throw toAppError(error, "Failed to create file signed url");
         };
     };
 };
@@ -70,10 +72,15 @@ export class DeleteFileFromS3UseCase {
 
     async execute(key: string): Promise<boolean> {
         try {
+            if (!key) {
+                throw new BadRequestError();
+            }
+
             const command = new DeleteObjectCommand({
                 Bucket: awsConfig.awsS3BucketName,
                 Key: key,
             });
+
             const res = await this.s3Client.send(command);
             if (res.$metadata.httpStatusCode === 200) {
                 await this.signedUrlService.delete(key);
@@ -81,9 +88,8 @@ export class DeleteFileFromS3UseCase {
             } else {
                 return false;
             };
-        } catch (error) {
-            log.error("DeleteFileFromS3UseCase failed", error as Error);
-            throw error;
+        } catch (error: unknown) {
+            throw toAppError(error, "Failed to delete file");
         };
     };
 };
